Add function 'insertIntoTargetModel()' to the genlib plugin
This is a generic function that inserts the contents of a library item
into a given target model. During this process, it calls
'transformLibraryItem()' on the library context so that special
implementations of the generic library can provide custom
implementations depending of the structure of their library item and of
the type of target model.